Rename files to correspond to classes names
